クラス名の変更と出力メソッドを別クラスへ切り出し
